# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GasForecast",
    "version": "1.0"
  },
  "paths": {
    "/api/Account/login": {
      "post": {
        "tags": [
          "Account"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginData"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginData"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Account/validate": {
      "get": {
        "tags": [
          "Account"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ElectricalUnitPassport/read_all": {
      "get": {
        "tags": [
          "ElectricalUnitPassport"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ElectricalUnitPassport"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ElectricalUnitPassport"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ElectricalUnitPassport"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ElectricalUnitPassport/{id}": {
      "get": {
        "tags": [
          "ElectricalUnitPassport"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ElectricalUnitPassport"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectricalUnitPassport"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectricalUnitPassport"
                }
              }
            }
          }
        }
      }
    },
    "/api/ElectricalUnitPassport/create": {
      "post": {
        "tags": [
          "ElectricalUnitPassport"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ElectricalUnitPassportCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ElectricalUnitPassportCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ElectricalUnitPassportCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ElectricalUnitPassport"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectricalUnitPassport"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectricalUnitPassport"
                }
              }
            }
          }
        }
      }
    },
    "/api/ElectricalUnitPassport/update/{id}": {
      "put": {
        "tags": [
          "ElectricalUnitPassport"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ElectricalUnitPassportUpdateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ElectricalUnitPassportUpdateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ElectricalUnitPassportUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ElectricalUnitPassport/delete/{id}": {
      "delete": {
        "tags": [
          "ElectricalUnitPassport"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ElectricityConsumptionCalculation/calculate": {
      "get": {
        "tags": [
          "ElectricityConsumptionCalculation"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ElectricityCalculationRequestDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ElectricityCalculationRequestDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ElectricityCalculationRequestDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ElectricityCalculationResponseDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ElectricityCalculationResponseDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ElectricityCalculationResponseDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ElectricityPowerStation/read_all": {
      "get": {
        "tags": [
          "ElectricityPowerStation"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ElectricalPowerStation"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ElectricalPowerStation"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ElectricalPowerStation"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ElectricityPowerStation/add": {
      "post": {
        "tags": [
          "ElectricityPowerStation"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ElectricalPowerStationCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ElectricalPowerStationCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ElectricalPowerStationCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ElectricalPowerStation"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectricalPowerStation"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectricalPowerStation"
                }
              }
            }
          }
        }
      }
    },
    "/api/ElectricityPowerStation/{id}": {
      "get": {
        "tags": [
          "ElectricityPowerStation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ElectricalPowerStation"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectricalPowerStation"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectricalPowerStation"
                }
              }
            }
          }
        }
      }
    },
    "/api/ElectricityPowerStation/delete/{id}": {
      "delete": {
        "tags": [
          "ElectricityPowerStation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ElectricityPowerStation/update/{id}": {
      "put": {
        "tags": [
          "ElectricityPowerStation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ElectricalPowerStationUpdateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ElectricalPowerStationUpdateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ElectricalPowerStationUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ElectricalPowerStation"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectricalPowerStation"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectricalPowerStation"
                }
              }
            }
          }
        }
      }
    },
    "/api/ElectricityPowerStation/get_passport_by_station_id/{id}": {
      "get": {
        "tags": [
          "ElectricityPowerStation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ElectricityPowerStation/stations_grouped_by_unit_type": {
      "get": {
        "tags": [
          "ElectricityPowerStation"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ElectricalPowerStation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "activeUnitsCount": {
            "type": "integer",
            "format": "int32"
          },
          "unitType": {
            "type": "string",
            "nullable": true
          },
          "launchDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ElectricalPowerStationCreateDTO": {
        "required": [
          "activeUnitsCount",
          "launchDate",
          "name",
          "unitType"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "unitType": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "activeUnitsCount": {
            "maximum": 30,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "launchDate": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
      "ElectricalPowerStationUpdateDTO": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "unitType": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "activeUnitsCount": {
            "maximum": 30,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "launchDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ElectricalUnitPassport": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "unitType": {
            "type": "string",
            "nullable": true
          },
          "engineType": {
            "type": "string",
            "nullable": true
          },
          "ratedPower": {
            "type": "integer",
            "format": "int32"
          },
          "standartPower": {
            "type": "integer",
            "format": "int32"
          },
          "consumptionNorm": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ElectricalUnitPassportCreateDTO": {
        "required": [
          "consumptionNorm",
          "engineType",
          "ratedPower",
          "standartPower",
          "unitType"
        ],
        "type": "object",
        "properties": {
          "unitType": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "engineType": {
            "minLength": 1,
            "pattern": "^(Поршневой|Газотурбинный)$",
            "type": "string"
          },
          "ratedPower": {
            "maximum": 100000,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "standartPower": {
            "maximum": 100000,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "consumptionNorm": {
            "maximum": 1000,
            "minimum": 0.1,
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ElectricalUnitPassportUpdateDTO": {
        "type": "object",
        "properties": {
          "unitType": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "engineType": {
            "pattern": "^(Поршневой|Газотурбинный)$",
            "type": "string",
            "nullable": true
          },
          "ratedPower": {
            "maximum": 100000,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "standartPower": {
            "maximum": 100000,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "consumptionNorm": {
            "maximum": 1000,
            "minimum": 0.1,
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ElectricityCalculationRequestDTO": {
        "required": [
          "lowerHeatingValue",
          "operatingHours",
          "outsideTemperature",
          "stationId",
          "unitPowerPercentage"
        ],
        "type": "object",
        "properties": {
          "stationId": {
            "type": "integer",
            "format": "int32"
          },
          "outsideTemperature": {
            "maximum": 40,
            "minimum": -30,
            "type": "number",
            "format": "double"
          },
          "operatingHours": {
            "maximum": 100000,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "unitPowerPercentage": {
            "maximum": 100,
            "minimum": 20,
            "type": "number",
            "format": "double"
          },
          "lowerHeatingValue": {
            "maximum": 10000,
            "minimum": 7000,
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ElectricityCalculationResponseDTO": {
        "type": "object",
        "properties": {
          "gasConsumption": {
            "type": "number",
            "format": "double"
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "calculationTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "LoginData": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}